# internal/handler/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Host API
  description: |
    Internal static website hosting. Every site is served at the root of its
    own host, `<sitename>.<owner>.<base>/` (for example
    `todo.alice.example.com/`). The owner host `<owner>.<base>/` serves the
    person's or team's index page. Until the owner's certificate is ready it
    also serves their sites at `<owner>.<base>/<sitename>/`; after that,
    those addresses, and `/api/sites/<sitename>/...` on the owner host,
    redirect to each site's own host with path and query kept (`301` for GET
    and HEAD, `308` otherwise). v1.2 addresses of the form
    `<owner>--<sitename>.<base>` only redirect. Always use the `url` from the
    latest response rather than composing an address. Uploads may
    be gzip-compressed tar or ZIP archives; nested regular-file downloads such
    as ZIP and DMG artifacts are supported.

    A site is owned by a person or a team. Team members change a team's
    sites through owner-qualified collaboration routes with optimistic
    `ETag` / `If-Match` protection.

    Every site has an access level: `only_me` (the owner, or the team's
    members; the default for a new site), `specific` (plus named viewers),
    `company` (anyone signed in with the link),
    `listed` (company, plus showcase and search), or `network` (anyone who
    can reach the server, no sign-in). `network` is requested with a reason
...
    network site's pages, assets and state, and change nothing.

...

...

...

...
servers:
...
paths:
...
  /api/sites/{sitename}/access:
    post:
      summary: Set who can open a site
      description: |
        Sets who can open the site. Any level except `network` takes effect at
        once; moving off `network` ends its approval, and moving to any level
        withdraws a pending network request. `network` needs a `reason` (at
        most 500 characters) and returns `202`: the site keeps its level until
        an admin approves the request. The site's address does not change;
        add viewers for `specific` through the viewers routes. While an admin's
        restriction stands (`access_decision` `restricted`), anything but
        `only_me` is refused with `409 site_restricted_by_admin`.
      tags: [Sites]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/sitename'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteAccessRequest'
      responses:
        '200':
          description: Access level changed
          headers:
            ETag:
              $ref: '#/components/headers/SiteETag'
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteAccessResponse'
        '202':
          description: Network access requested; the site keeps its current level until an admin approves
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteAccessResponse'
        '400':
          $ref: '#/components/responses/ManagementBadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: '`already_network`: the site is already open to the network; `site_restricted_by_admin`: an admin restricted the site, so only `only_me` is accepted until an admin lifts it (the body carries the admin''s `reason`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/collaboration/sites/{owner}/{sitename}/versions/{version}/preview:
    get:
      summary: Get a preview link for one kept version
      description: |
        Owner/member operation. Returns a private address that shows the
        version (one stored with `publish=false`, or an older one) without
        changing what visitors see. It lasts one hour and opens only for the
        site's owner or a member of the owning team with their own sign-in,
        whatever the site's access level; it is not indexed or cached, and
        saves and uploads from it are refused (403 `preview_read_only`). Make
        it live with the rollback route. A publish-scope key may call it.
      tags: [Collaboration]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/owner'
        - $ref: '#/components/parameters/sitename'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: The preview link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preview'
        '400':
          $ref: '#/components/responses/SkillVersionRequired'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          description: Previews are not available on this server
...
  /api/collaboration/sites/{owner}/{sitename}/access:
    post:
      summary: Set who can open an owner-qualified site
      description: |
        Owner/member operation. Same as `POST /api/sites/{sitename}/access`, for a
        site in a team's namespace.
        Sets who can open the site. Any level except `network` takes effect at
        once; moving off `network` ends its approval, and moving to any level
        withdraws a pending network request. `network` needs a `reason` (at
        most 500 characters) and returns `202`: the site keeps its level until
        an admin approves the request. The site's address does not change;
        add viewers for `specific` through the viewers routes. While an admin's
        restriction stands (`access_decision` `restricted`), anything but
        `only_me` is refused with `409 site_restricted_by_admin`.
      tags: [Collaboration]
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/SimpleHostClient'
        - $ref: '#/components/parameters/owner'
        - $ref: '#/components/parameters/sitename'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteAccessRequest'
      responses:
        '200':
          description: Access level changed
          headers:
            ETag:
              $ref: '#/components/headers/SiteETag'
            X-Simple-Host-Latest-Skill-Version:
              $ref: '#/components/headers/LatestSkillVersion'
            X-Simple-Host-Minimum-Skill-Version:
              $ref: '#/components/headers/MinimumSkillVersion'
            X-Simple-Host-Update-URL:
              $ref: '#/components/headers/SkillUpdateURL'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteAccessResponse'
        '202':
          description: Network access requested; the site keeps its current level until an admin approves
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteAccessResponse'
        '400':
          $ref: '#/components/responses/ManagementBadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: '`already_network`: the site is already open to the network; `site_restricted_by_admin`: an admin restricted the site, so only `only_me` is accepted until an admin lifts it (the body carries the admin''s `reason`)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/access:
    get:
      summary: Read a site's access log
      description: |
        A non-admin must pass `owner` (their own username or a team they are
        in). What comes back depends on `ACCESS_LOG_VISIBILITY`: under
        `counts` (the default) a non-admin gets `AccessCounts` — views per
        day and distinct viewers, the last 30 days unless `from`/`to` say
        otherwise, never who; under `owner` they get rows with user ids but
        no IP or user agent, or `AccessCounts` with `summary=counts`; under
        `admin` they are refused. Admins always
        get full rows. API-key callers send `X-Skill-Version` as on the other
        management routes.
      tags: [Audit]
      security:
        - apiKey: []
      parameters:
        - name: owner
          in: query
          schema:
            type: string
        - name: site
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: cursor
          in: query
          schema:
            type: string
        - name: summary
          in: query
          description: '`counts` asks a non-admin for `AccessCounts` whatever the visibility (short of `admin`).'
          schema:
            type: string
            enum: [counts]
      responses:
        '200':
          description: Aggregates under `counts` (or with `summary=counts`) for a non-admin; otherwise a page of rows
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AccessCounts'
                  - type: object
                    properties:
                      entries:
                        type: array
                        items:
                          type: object
                      next_cursor:
                        type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  headers:
    LatestSkillVersion:
      description: Latest skill bundle version published by this server.
      schema:
        type: string
        example: 0.8.1
    MinimumSkillVersion:
      description: Oldest skill version accepted for API-key management calls.
      schema:
        type: string
        example: 0.8.0
    SkillUpdateURL:
      description: Same-origin page describing the permissioned update flow.
      schema:
        type: string
        example: /install.html
...
    SiteETag:
      description: Opaque validator containing the immutable site incarnation and active version.
      schema:
        type: string
        example: '"site-4b256b3b-9d15-4ce5-bba7-b04ef26f49e5-v7"'
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        An API key, for CI and other automation; agents normally connect
        through `/mcp` with the company sign-in instead. Each key has a scope
        chosen when it is created and an expiry (at most 365 days).
        `publish` (the default) deploys, updates and rolls back sites, reads
        versions, saved data and assets, and calls `/api/me` and `/mcp`.
        `full` adds everything else the person can do (deleting sites, access
        levels, viewers, teams, the audit and access logs, search), except
        administration. `offboard` (admins only) calls
        `POST /api/admin/users/disable` and nothing else. No key reaches any
        other `/api/admin/*` route. OAuth connector tokens work only on
        `/mcp`.
  parameters:
    SkillVersion:
      name: X-Skill-Version
      in: header
      required: false
      description: Required on API-key management requests made by an agent. Strict numeric major.minor.patch; the server rejects malformed or below-minimum values.
      schema:
        type: string
        pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)$'
      example: 0.8.1
    SimpleHostClient:
      name: X-Simple-Host-Client
      in: header
      required: false
      description: Intentional non-skill client classification. Direct API integrations send `api`; `control-ui` is reserved for the first-party account page. This is compatibility metadata, not authentication, and never replaces X-API-Key.
      schema:
        type: string
        enum: [api, control-ui]
      example: api
    sitename:
      name: sitename
      in: path
      required: true
      schema:
        type: string
      example: my-cool-site
...
    owner:
      name: owner
      in: path
      required: true
      description: The site's owner — a username or a stored team name (`team-<name>`), as list responses give it.
      schema:
        type: string
      example: alice.smith
    version:
      name: version
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      example: 4
...
  schemas:
...
    CollaborationSite:
      type: object
      description: One canonical site the caller owns or whose team the caller is in. It contains no API keys.
      required: [id, name, owner_username, access_role, active_version, public, access, public_path, url, etag, created_at, updated_at, analytics]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner_username:
          type: string
        access_role:
          type: string
          enum: [owner, member, viewer]
          description: '`viewer` appears only with `include=shared`: the site is shared with the caller, who can open it and nothing more.'
        shared_via:
          type: string
          description: On a viewer entry only, the team whose viewer grant lets the caller open the site; empty when the grant names the caller.
        active_version:
          type: integer
          minimum: 1
        public:
          type: boolean
          description: True when the access level is `listed` or `network`.
        access:
          type: string
          enum: [only_me, specific, company, listed, network]
          description: Who can open the site.
        network_request:
          type: object
          description: Present while a request to open the site to the network waits for an admin.
          properties:
            status:
              type: string
              enum: [pending]
            reason:
              type: string
            requested_at:
              type: string
              format: date-time
            approvals:
              type: integer
              description: Admin approvals so far.
            approvals_required:
              type: integer
              description: Approvals needed (`NETWORK_ACCESS_APPROVALS`, 1 or 2).
        access_decision:
          type: object
          description: The last admin decision about who can open the site. `declined` and `revoked` stay until the next network request; `restricted` until an admin lifts it (the owner cannot).
          required: [decision, at]
          properties:
            decision:
              type: string
              enum: [declined, revoked, restricted]
            at:
              type: string
              format: date-time
            reason:
              type: string
              description: The admin's note; always present for `restricted`.
        public_path:
          type: string
          example: https://my-cool-site.alice-smith.simple-host.example.com/
        url:
          type: string
          example: https://my-cool-site.alice-smith.simple-host.example.com/
        etag:
          type: string
          example: '"site-4b256b3b-9d15-4ce5-bba7-b04ef26f49e5-v7"'
        new_version:
          type: integer
          description: Only after a deploy with `publish=false`, the version it stored, which is not live.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        analytics:
          $ref: '#/components/schemas/SiteAnalytics'
    Preview:
      type: object
      required: [url, expires_at, version, live]
      properties:
        url:
          type: string
          description: The private preview address, the site's own address plus `_preview/<token>/`. It opens only for the site's owner or a member of the owning team, signed in.
          example: https://my-cool-site.alice-smith.simple-host.example.com/_preview/5-1790000000-Qm9n/
        expires_at:
          type: string
          format: date-time
          description: One hour from when it was made; after that the address answers 410.
        version:
          type: integer
        live:
          type: boolean
          description: Whether that version is already live.
...
    SiteAccessRequest:
      type: object
      required: [level]
      properties:
        level:
          type: string
          enum: [only_me, specific, company, listed, network]
        reason:
          type: string
          maxLength: 500
          description: Required for `network`; the admin reads it.
    SiteAccessResponse:
      allOf:
        - $ref: '#/components/schemas/CollaborationSite'
        - type: object
          properties:
            note:
              type: string
...
    AccessCounts:
      type: object
      required: [from, to, unique_viewers, days]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        unique_viewers:
          type: integer
          format: int64
        days:
          type: array
          items:
            type: object
            required: [day, views, unique_viewers]
            properties:
              day:
                type: string
                format: date
              views:
                type: integer
                format: int64
              unique_viewers:
                type: integer
                format: int64
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        reason:
          type: string
          description: With `site_restricted_by_admin`, the admin's note.
...
  responses:
...
    SkillVersionRequired:
      description: A valid API key was supplied, but the management client omitted a supported skill version or explicit non-skill classification.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Simple-Host-Latest-Skill-Version:
          $ref: '#/components/headers/LatestSkillVersion'
        X-Simple-Host-Minimum-Skill-Version:
          $ref: '#/components/headers/MinimumSkillVersion'
        X-Simple-Host-Update-URL:
          $ref: '#/components/headers/SkillUpdateURL'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SkillVersionRequired'
    ManagementBadRequest:
      description: Invalid management request, or a valid API-key client that omitted a supported skill version or explicit non-skill classification.
      headers:
        Cache-Control:
          description: "no-store on skill-version bootstrap errors; otherwise handler-specific."
          schema:
            type: string
        X-Simple-Host-Latest-Skill-Version:
          $ref: '#/components/headers/LatestSkillVersion'
        X-Simple-Host-Minimum-Skill-Version:
          $ref: '#/components/headers/MinimumSkillVersion'
        X-Simple-Host-Update-URL:
          $ref: '#/components/headers/SkillUpdateURL'
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Error'
              - $ref: '#/components/schemas/SkillVersionRequired'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 'Missing or invalid credentials. For an `X-API-Key` that did not authenticate, `code` says why: `key_expired` (the body names the date; `expired_at` carries it), `key_revoked`, `key_owner_disabled` or `key_not_recognised`. A key with 14 days or less left gets `X-Key-Expires` (RFC 3339) and `X-Simple-Host-Notice` on every response.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated caller does not have the owner-only permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Site, retained version, or membership was not found or is not accessible
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: Process-local rate limit exceeded; wait at least the advertised number of seconds before retrying
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'